flask/requirements.txt
Flask
Flask-SQLAlchemy
Flask-JWT-Extended
Flask-Bcrypt
Flask-CORS
PyMySQL
cryptography
python-dotenv
